+++ corpora/vendored/other/multiline-string-trailing-space-out.hs
@@ -1,0 +1,9 @@
+{-# LANGUAGE MultilineStrings #-}
+
+script :: String
+script =
+  """
+  (function (x) { 
+    log(x); 
+  })
+  """